refactor(app_view): split session/window scaffolding into session.rs
Lifts the small leaf concerns out of mod.rs:

  * `SessionBundle` — the moved-between-windows live-session
    container.
  * `WindowInit` — the three-variant discriminator passed to
    `open_app_window` (Fresh / WithSession / FreshAutoConnect).
  * `geometry_to_bounds` / `bounds_to_geometry` — the
    `WindowGeometry` ↔ `gpui::Bounds<Pixels>` helpers used by
    `open_app_window`'s saved-bounds restore + the
    `on_window_should_close` snapshot path.

Behaviour-preserving. `SessionBundle`'s fields stay package-
private to the `app_view` module via `pub(super)` so the parent's
`extract_session` / `install_session` methods continue to
construct and destructure it directly — the bundle is
intentionally a thin movable container, not a closed type with
constructors.

While I was here: reattached the orphan doc comment for
`open_app_window`, which had drifted into the gap between
`WindowInit` and the geometry helpers when those helpers were
inserted in an earlier change. It now sits where it belongs.

Verified: `cargo clippy --all-targets -- -D warnings` and
`cargo test` both pass (63 tests, unchanged).
